Move types out of example data
