+++ src/IsaacApiTypes.tsx
@@ -96,6 +96,21 @@ export interface IsaacAnvilQuestionDTO extends QuestionDTO {
 export interface IsaacConceptPageDTO extends SeguePageDTO {
 }
 
+export interface BookSectionDTO extends ContentDTO {
+    label?: string;
+    pageId?: string;
+}
+
+export interface BookChapterDTO extends ContentDTO {
+    label?: string;
+    sections?: BookSectionDTO[];
+}
+
+export interface IsaacBookIndexPageDTO extends SeguePageDTO {
+    chapters?: BookChapterDTO[];
+    coverImage?: ImageDTO;
+}
+
 export interface IsaacEventPageDTO extends ContentDTO {
     date?: Date;
     bookingDeadline?: Date;
